Refactoring with claude

+++ BerlinSand/run_all.py
@@ -1,8 +1,16 @@
+#-------------------------------------------------------------------------------
+# Name:        run_all
+# Purpose:     drive a ZSoil analysis end to end: run the solver on the
+#              Berlin Sand excavation benchmark, then call each plot_*.py
+#              module to turn the result files into figures
+#
+# Author:      Matthias Preisig
+# Created:     2026
+# Copyright:   (c) GeoDev Sarl
+#-------------------------------------------------------------------------------
 import os
 import shutil
 import subprocess
-import numpy as np
-
 
 import plot_excavation2D
 import plot_interfaces
@@ -10,40 +18,39 @@ import plot_time_histories
 import plot_nodal_maps
 import plot_element_maps
 
+from demo_config import PROBLEM, STEPS_TO_PLOT, STEP_LABELS
 
-# Config section
 
-prob = r'HS-Brick-Exc-Berlin-Sand-2phase_2026'
-
+# ------------------------- Config section -------------------------
+# Edit these two for your own machine/problem. STEPS_TO_PLOT / STEP_LABELS
+# come from demo_config.py, which the plot_*.py scripts also read as their
+# own command-line defaults.
 ZSoil_exe = r'C:/Program Files/ZSoil/ZSoil 2026 v26.03/Z_Soil.exe'
 working_dir = os.path.realpath('./res-files')
+# ----------------------- End of config section ----------------------
 
-# End of config section
+if not os.path.isfile(ZSoil_exe):
+    raise FileNotFoundError(
+        "ZSoil executable not found at '%s' - edit ZSoil_exe at the top of "
+        "run_all.py to point at your installation." % ZSoil_exe
+    )
 
-os.mkdir(working_dir)
-shutil.copy('./inp-files/%s.inp'%(prob),'%s/%s.inp'%(working_dir,prob))
+os.makedirs(working_dir, exist_ok=True)
+shutil.copy('./inp-files/%s.inp' % PROBLEM, '%s/%s.inp' % (working_dir, PROBLEM))
 
-subprocess.check_call(
-        '%s "%s\\%s" /S' % (ZSoil_exe, working_dir, prob+'.inp'),
+try:
+    subprocess.check_call(
+        '%s "%s\\%s" /S' % (ZSoil_exe, working_dir, PROBLEM + '.inp'),
         cwd=working_dir,
     )
 
-
-step_labels = {
-    3:'Exc. at -4.8 m',
-    4:'Anchor level 1',
-    5:'Exc. at -9.3 m',
-    6:'Anchor level 2',
-    7:'Exc. at -14.35 m',
-    8:'Anchor level 3',
-    9:'Exc. at -16.8 m'
-    }
-steps_to_plot = [3,5,7,9]
-
-plot_excavation2D.main(working_dir, prob, steps_to_plot, step_labels)
-plot_interfaces.main(working_dir, prob, steps_to_plot, step_labels)
-plot_time_histories.main(working_dir, prob)
-plot_nodal_maps.main(working_dir, prob, steps_to_plot[-1])
-plot_element_maps.main(working_dir, prob, steps_to_plot[-1])
-
-shutil.rmtree(working_dir)
+    plot_excavation2D.main(working_dir, PROBLEM, STEPS_TO_PLOT, STEP_LABELS)
+    plot_interfaces.main(working_dir, PROBLEM, STEPS_TO_PLOT, STEP_LABELS)
+    plot_time_histories.main(working_dir, PROBLEM)
+    plot_nodal_maps.main(working_dir, PROBLEM, STEPS_TO_PLOT[-1])
+    plot_element_maps.main(working_dir, PROBLEM, STEPS_TO_PLOT[-1])
+except Exception:
+    print("run_all.py: an error occurred, leaving '%s' in place for inspection." % working_dir)
+    raise
+else:
+    shutil.rmtree(working_dir)
